message when no /dev/kmsg
