+++ gprofiler/kernel_messages.py
@@ -1,45 +1,19 @@
-import os
-import time
-import errno
-
-from typing import List, Tuple
 from gprofiler.log import get_logger_adapter
+from gprofiler.utils import get_kernel_release
 
 logger = get_logger_adapter(__name__)
-
-# See linux/printk.h
-CONSOLE_EXT_LOG_MAX = 8192
+kernel_release = get_kernel_release()
 
 
-class DevKmsgReader:
-    # The /dev/kmsg interfaced is described at Documentation/ABI/testing/dev-kmsg in the kernel source tree
-    # and can be viewed at https://github.com/torvalds/linux/blob/master/Documentation/ABI/testing/dev-kmsg.
+class DummyMessageReader:
     def __init__(self):
-        self.dev_kmsg_fd = os.open("/dev/kmsg", os.O_RDONLY)
-        os.set_blocking(self.dev_kmsg_fd, False)
-        # skip all historical messages:
-        os.lseek(self.dev_kmsg_fd, 0, os.SEEK_END)
+        print("This kernel does not support the new /dev/kmsg interface for reading messages.")
+        print("Profilee error monitoring not available.")
+        print()
+        logger.warning("Profilee error monitoring not available.")
 
     def iter_new_messages(self):
-        messages: List[Tuple[float, bytes]] = []
-        try:
-            # Each read() is one message
-            while True:
-                message = os.read(self.dev_kmsg_fd, CONSOLE_EXT_LOG_MAX)
-                messages.append((time.time(), message))
-        except OSError as e:
-            if e.errno != errno.EAGAIN:
-                raise
-
-        yield from self._parse_raw_messages(messages)
-
-    @staticmethod
-    def _parse_raw_messages(messages: List[Tuple[float, bytes]]):
-        for timestamp, message in messages:
-            prefix, text = message.decode().split(";", maxsplit=1)
-            fields = prefix.split(",")
-            level = int(fields[0])
-            yield timestamp, level, text
+        return []
 
 
 class KernelMessagePublisher:
@@ -57,3 +31,11 @@ class KernelMessagePublisher:
 
     def subscribe(self, callback):
         self.subscribers.append(callback)
+
+
+if kernel_release >= (3, 5):
+    from gprofiler.devkmsg import DevKmsgReader
+
+    DefaultMessageReader = DevKmsgReader
+else:
+    DefaultMessageReader = DummyMessageReader
