+++ lint.sh
@@ -10,7 +10,7 @@ if [[ "$1" = "--ci" ]]; then
     check_arg="--check"
 fi
 
-isort --settings-path .isort.cfg -rc .
+isort --settings-path .isort.cfg .
 black --line-length 120 $check_arg .
 flake8 --config .flake8 .
 mypy .
